# api.yaml
swagger: "2.0"
info:
  title: "DeepSpace Docs"
  description: >
    Docs de DeepSpace API.
  version: "1.0.0"
host: "${DB_HOST}:${PORT}"
schemes:
  - "http"
...
paths:
  /users:
    get:
      tags: ["users", ]
      summary: "Recupera todos los usuarios"
      description: "Recupera todos los usuarios de la base de datos."
      responses:
        200:
          description: "Una lista de usuarios"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
        500:
          description: "Internal Server Error"
          schema:
            $ref: '#/definitions/Error'
    post:
      tags: ["users", ]
      summary: "Crear un nuevo usuario"
      description: "Crea un nuevo usuario con los datos proporcionados."
      parameters:
        - in: "body"
          name: "body"
          description: "El usuario que debe ser creado"
          required: true
          schema:
            $ref: "#/definitions/UserEntity"
      responses:
        201:
          description: "Usuario creado correctamente."
          schema:
            $ref: "#/definitions/User"
        500:
          description: "Internal Server Error"
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  Error:
    type: "object"
    properties:
      error:
        type: "string"
...
  User:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      lastName:
        type: "string"
      profilePictureUrl:
        type: "string"
  UserEntity:
    type: "object"
    properties:
      name:
        type: "string"
      last_name:
        type: "string"
      password:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      profile_picture:
        type: "string"        
...